refactor(ranking): add RankingInfo DTO and improve MyRankingResponse - Add RankingInfo record for position and total players - Add JsonProperty annotations to MyRankingResponse for API consistency

+++ src/main/java/dev/gamified/GamifiedPlatform/dtos/response/ranking/MyRankingResponse.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 @Builder
 public record MyRankingResponse(
         @JsonProperty("position")
-        Integer position,
+        Long position,
 
         @JsonProperty("total_players")
         Long totalPlayers,
