+++ src/main/java/dev/gamified/GamifiedPlatform/dtos/response/ranking/RankingInfo.java
@@ -1,7 +1,7 @@
 package dev.gamified.GamifiedPlatform.dtos.response.ranking;
 
 public record RankingInfo(
-        Integer position,
+        Long position,
         Long totalPlayers
 ) {
 }
